requirements.txt
fastmcp
langchain
langchain-groq
langchain-community
python-dotenv
langchain-mcp-adapters